fixed various issues such as, int_t != size_t for C, array continuguity forcing for numpy, turn off bounds checks. More to fix, especially outside of eval_conc

+++ pyjacob.cuh
@@ -3,11 +3,11 @@
 #ifndef CU_PYJAC_HEAD
 #define CU_PYJAC_HEAD
 
-void cu_eval_conc (const size_t num, const double* T, const double* pres, const double* mass_frac, double* mw_avg, double* rho, double* conc);
-void cu_eval_rxn_rates (const size_t num, const double* T, const double* pres, const double * C, double * fwd_rxn_rates, double * rev_rxn_rates);
-void cu_get_rxn_pres_mod (const size_t num, const double* T, const double* pres, const double * C, double * pres_mod);
-void cu_eval_spec_rates (const size_t num, const double* fwd_rates, const double* rev_rates, const double* pres_mod, double* spec_rates);
-void cu_dydt (const size_t num, const double* pres, const double* y, double* dy);
-void cu_eval_jacob (const size_t num, const double* pres, const double* y, double* jac);
+void cu_eval_conc (const int num, const double* T, const double* pres, const double* mass_frac, double* mw_avg, double* rho, double* conc);
+void cu_eval_rxn_rates (const int num, const double* T, const double* pres, const double * C, double * fwd_rxn_rates, double * rev_rxn_rates);
+void cu_get_rxn_pres_mod (const int num, const double* T, const double* pres, const double * C, double * pres_mod);
+void cu_eval_spec_rates (const int num, const double* fwd_rates, const double* rev_rates, const double* pres_mod, double* spec_rates);
+void cu_dydt (const int num, const double* pres, const double* y, double* dy);
+void cu_eval_jacob (const int num, const double* pres, const double* y, double* jac);
 
 #endif
